requirements.txt
python-telegram-bot==21.6
supabase==2.9.1
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.2
fastapi==0.115.0
uvicorn==0.30.0
